% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{abm.ees2}
\alias{abm.ees2}
\title{EvoPER Evolutionary Strategy 2}
\usage{
abm.ees2(objective, options = NULL)
}
\arguments{
\item{objective}{An instance of ObjectiveFunction (or subclass) class \link{ObjectiveFunction}}

\item{options}{An apropiate instance from a sublclass of \link{Options} class}
}
\description{
This function tries to provide a rough approximation to
best solution when no information is available for the correct range
of input parameters for the objective function. It can useful for
studying the behavior of individual-based models with high
variability in the output variables showing nonlinear behaviors.
}

