% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{abm.sda}
\alias{abm.sda}
\title{Simulated Dilution Approximation}
\usage{
abm.sda(objective, options = NULL)
}
\arguments{
\item{objective}{An instance of ObjectiveFunction (or subclass) class \link{ObjectiveFunction}}

\item{options}{An apropiate instance from a sublclass of \link{Options} class}
}
\description{
This function tries to provide a better approximation to the
best solution when no information is available on the correct range of
input parameters for the objective function. Basically the function search
for the best cost n replications and adjust the range based on that value
of best particles and a 10^-1 of initially provided range.
}

