% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-acor.R
\name{acor.archive}
\alias{acor.archive}
\title{acor.archive}
\usage{
acor.archive(s, f, w, k, T = NULL)
}
\arguments{
\item{s}{The solution 'ants'}

\item{f}{The evaluation of solution}

\item{w}{The weight vector}

\item{k}{The archive size}

\item{T}{The current archive}
}
\value{
The solution archive
}
\description{
This function is used for creating and maintaining the ACOr archive 'T'.
The function keeps the track of 'k' solotion in the archive.
}
\references{
[1] Socha, K., & Dorigo, M. (2008). Ant colony optimization for continuous domains.
European Journal of Operational Research, 185(3), 1155-1173.
http://doi.org/10.1016/j.ejor.2006.06.046
}
