% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-acor.R
\name{acor.F}
\alias{acor.F}
\title{acor.F}
\usage{
acor.F(T)
}
\arguments{
\item{T}{The solution archive}
}
\value{
The F matrix
}
\description{
Helper function for extracting the 'F' function evaluations
from archive ACOr 'T'
}

