% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{acor.lthgaussian}
\alias{acor.lthgaussian}
\title{Select the lth gaussian function}
\usage{
acor.lthgaussian(W)
}
\arguments{
\item{W}{The vector of weights}
}
\value{
The index of lht gaussian function
}
\description{
Given a weight vector calculate the probabilities of selecting
the lth gaussian function and return the index of lht gaussian selected with
probability p
}
\references{
[1] Socha, K., & Dorigo, M. (2008). Ant colony optimization for continuous domains.
European Journal of Operational Research, 185(3), 1155-1173.
http://doi.org/10.1016/j.ejor.2006.06.046
}

