% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-acor.R
\name{acor.N}
\alias{acor.N}
\title{acor.N}
\usage{
acor.N(T)
}
\arguments{
\item{T}{The solution archive}
}
\value{
The size 'n' of a solution 's'
}
\description{
Helper function for getting the size of solution
}

