% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{acor.probabilities}
\alias{acor.probabilities}
\title{Gaussian kernel choosing probability}
\usage{
acor.probabilities(W, l = NULL)
}
\arguments{
\item{W}{The vector of weights}

\item{l}{The lth element of algorithm solution archive T}
}
\value{
The vector of probabilities 'p'
}
\description{
Calculate the probability of choosing the lth Gaussian function
}
\references{
[1] Socha, K., & Dorigo, M. (2008). Ant colony optimization for continuous domains.
European Journal of Operational Research, 185(3), 1155-1173.
http://doi.org/10.1016/j.ejor.2006.06.046
}

