% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-acor.R
\name{acor.S}
\alias{acor.S}
\title{acor.S}
\usage{
acor.S(T)
}
\arguments{
\item{T}{The solution archive}
}
\value{
The solution matrix
}
\description{
Helper function for extracting solution 'S' from archive 'T'
}
