% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-acor.R
\name{acor.sigma}
\alias{acor.sigma}
\title{Sigma calculation for ACOr}
\usage{
acor.sigma(Xi, k, T)
}
\arguments{
\item{Xi}{The algorithm parameter}

\item{k}{The solution archive size}

\item{T}{The solution archive}
}
\value{
The sigma value
}
\description{
Calculate the value of sigma
}
\references{
[1] Socha, K., & Dorigo, M. (2008). Ant colony optimization for continuous domains.
European Journal of Operational Research, 185(3), 1155-1173.
http://doi.org/10.1016/j.ejor.2006.06.046
}
