% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-acor.R
\name{acor.updateants}
\alias{acor.updateants}
\title{acor.updateants}
\usage{
acor.updateants(S, N, W, t.mu, t.sigma)
}
\arguments{
\item{S}{The current solution ants}

\item{N}{The numnber of required ants in solution}

\item{W}{The weight vector}

\item{t.mu}{The 'mean' from solution archive}

\item{t.sigma}{The value of sigma from solution archive}
}
\value{
The new solution ants
}
\description{
Update the solution using the gaussian kernel
}
\references{
[1] Socha, K., & Dorigo, M. (2008). Ant colony optimization for continuous domains.
European Journal of Operational Research, 185(3), 1155-1173.
http://doi.org/10.1016/j.ejor.2006.06.046
}

