% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{acor.W}
\alias{acor.W}
\title{acor.W}
\usage{
acor.W(T)
}
\arguments{
\item{T}{The solution archive}
}
\value{
The weight vector
}
\description{
Helper function for extracting the 'W' function evaluations
from archive ACOr 'T'
}

