% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{assert}
\alias{assert}
\title{assert}
\usage{
assert(expresion, string)
}
\arguments{
\item{expresion}{Some logical expression}

\item{string}{The text message to show if expression does not hold}
}
\description{
The assert function stop the execution if the logical
expression given by the parameter \code{expresion} is false.
}
