% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{basic_config}
\alias{basic_config}
\title{basic_config}
\usage{
basic_config(level = "INFO", file = NULL)
}
\arguments{
\item{level}{The log level (ERROR|WARN|INFO|DEBUG)}

\item{file}{The log file destination}
}
\description{
Wrapper for logging error messages.
}
