% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{bestFitness}
\alias{bestFitness}
\title{bestFitness}
\usage{
bestFitness(S)
}
\arguments{
\item{S}{The solution set}
}
\value{
The best fitness value
}
\description{
Given a set S of N solutions created with sortSolution, this function
returns the fitness component fot the best solution.
}
