% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{bestSolution}
\alias{bestSolution}
\title{bestSolution}
\usage{
bestSolution(S)
}
\arguments{
\item{S}{The solution set}
}
\value{
The best solution
}
\description{
Given a set S of N solutions created with sortSolution, this function
returns the best solution found.
}
