% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{cbuf}
\alias{cbuf}
\title{cbuf}
\usage{
cbuf(b, v, e)
}
\arguments{
\item{b}{The variable holding the current buffer content}

\item{v}{The new valued to be added to b}

\item{e}{The length of circular buffer}
}
\value{
The buffer b plus the element v minus the least recently added element
}
\description{
Simple implementation of a circular buffer.
}

