% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{compare.algorithms1}
\alias{compare.algorithms1}
\title{compare.algorithms1}
\usage{
compare.algorithms1(F, seeds = c(27, 2718282, 36190727, 3141593, -91190721,
  -140743, 1321))
}
\arguments{
\item{F}{The function to be tested}

\item{seeds}{The random seeds which will be used for testing algorithms}
}
\description{
Compare the number of function evalutions and convergence for the
following optimization algorithms, ("saa","pso","acor","ees1").
}
\examples{
\dontrun{
 rm(list=ls())
 d.cigar4<- compare.algorithms1(f0.cigar4)
 d.schaffer4<- compare.algorithms1(f0.schaffer4)
 d.griewank4<- compare.algorithms1(f0.griewank4)
 d.bohachevsky4<- compare.algorithms1(f0.bohachevsky4)
 d.rosenbrock4<- compare.algorithms1(f0.rosenbrock4)
}

}
