% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{contourplothelper}
\alias{contourplothelper}
\title{contourplothelper}
\usage{
contourplothelper(d, x, y, z)
}
\arguments{
\item{d}{A data frame.}

\item{x}{A string with the dataframe column name for x axis}

\item{y}{A string with the dataframe column name for y axis}

\item{z}{A string with the dataframe column name for z axis}
}
\description{
Simple helper for ploting histograms
}

