% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-ees1.R
\name{ees1.challenge}
\alias{ees1.challenge}
\title{ees1.challenge}
\usage{
ees1.challenge(solution, objective)
}
\arguments{
\item{solution}{The Problem solution}

\item{objective}{The objective function}
}
\description{
Repeat the evalution of best solution to tacke with
variability.
}
