% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-ees1.R
\name{ees1.mating1}
\alias{ees1.mating1}
\title{ees1.mating1}
\usage{
ees1.mating1(solution, mu)
}
\arguments{
\item{solution}{The Problem solution}

\item{mu}{The mixing intensity ratio, from 0 to 1. The mix intensity
controls de the probability of chosing a worst solutions}
}
\description{
This function 'mix' the elements present in the solution. The
parameter 'mu' controls the intensity of mixing. Low values give preference
to best solution components and high values make the values being select randomly.
}
