% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{ees1.mutation}
\alias{ees1.mutation}
\title{ees1.mutation}
\usage{
ees1.mutation(solution, mates)
}
\arguments{
\item{solution}{The Problem solution}

\item{mates}{The mixed parents}
}
\description{
Performs the mutation on generated solution
}

