% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{ees1.recombination}
\alias{ees1.recombination}
\title{ees1.recombination}
\usage{
ees1.recombination(solution, mates)
}
\arguments{
\item{solution}{The Problem solution}

\item{mates}{The mixed parents}
}
\description{
Performs the recombination on solution
}

