% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{elog.debug}
\alias{elog.debug}
\title{elog.debug}
\usage{
elog.debug(...)
}
\arguments{
\item{...}{Variable number of arguments including a format string.}
}
\description{
Wrapper for logging debug messages.
}
