% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{elog.error}
\alias{elog.error}
\title{elog.error}
\usage{
elog.error(...)
}
\arguments{
\item{...}{Variable number of arguments including a format string.}
}
\description{
Wrapper for logging error messages.
}

