% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{elog.level}
\alias{elog.level}
\title{elog.level}
\usage{
elog.level(level = NULL)
}
\arguments{
\item{level}{The log level (ERROR|WARN|INFO|DEBUG)}
}
\value{
The log level
}
\description{
Configure the current log level
}

