% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{enforceBounds}
\alias{enforceBounds}
\title{enforceBounds}
\usage{
enforceBounds(particles, factors)
}
\arguments{
\item{particles}{The particle set}

\item{factors}{the defined range for objective function parameters}
}
\value{
The particle inside the valid limits
}
\description{
Checks if parameters fall within upper an lower bounds
}
