% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-ees1.R
\name{es.evaluate}
\alias{es.evaluate}
\title{es.evaluate}
\usage{
es.evaluate(f, s, enforce = TRUE)
}
\arguments{
\item{f}{A reference to an instance of objective function}

\item{s}{The set of solutions}

\item{enforce}{If true the values are enforced to fall within provided range}
}
\value{
The solution ordered by its fitness.
}
\description{
For each element in solution 's' evaluate the respective
fitness.
}
