% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{extremize}
\alias{extremize}
\title{extremize}
\usage{
extremize(type, objective, options = NULL)
}
\arguments{
\item{type}{The optimization method (aco,pso,saa,sda)}

\item{objective}{An instance of ObjectiveFunction (or subclass) class \link{ObjectiveFunction}}

\item{options}{An apropiate instance from a sublclass of \link{Options} class}
}
\description{
Entry point for optimization functions
}
\examples{
\dontrun{
 f<- PlainFunction$new(f0.rosenbrock2)

 f$Parameter(name="x1",min=-100,max=100)
 f$Parameter(name="x2",min=-100,max=100)

 extremize("pso", f)
}

}

