% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{f0.ackley}
\alias{f0.ackley}
\title{f0.ackley}
\usage{
f0.ackley(...)
}
\arguments{
\item{...}{The variadic list of function variables.}
}
\value{
The function value
}
\description{
The ackley function of N variables for testing
optimization methods. The global optima for the function is given by
xi = 0, forall i E {1...N}, f(x) = 0.
Domain xi E [-32.768, 32.768], for all i = 1, ..., d
}
\references{
https://www.sfu.ca/~ssurjano/ackley.html
}
