% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{f0.bohachevsky}
\alias{f0.bohachevsky}
\title{f0.bohachevsky}
\usage{
f0.bohachevsky(...)
}
\arguments{
\item{...}{The variadic list of function variables.}
}
\value{
The function value
}
\description{
The Bohachevsky function of N variables for testing
optimization methods. The global optima for the function is given by
xi = 0, forall i E {1...N}, f(x) = 0.
}
\references{
http://deap.gel.ulaval.ca/doc/dev/api/benchmarks.html
}
