% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{f0.bohachevsky4}
\alias{f0.bohachevsky4}
\title{f0.bohachevsky4}
\usage{
f0.bohachevsky4(x1, x2, x3, x4)
}
\arguments{
\item{x1}{The first function variable}

\item{x2}{The second function variable}

\item{x3}{The third function variable}

\item{x4}{The fourth function variable}
}
\value{
The function value
}
\description{
The Bohachevsky function of four variables for testing optimization
methods. The global optima for the function is given by
xi = 0, forall i E {1...N}, f(x) = 0.
}

