% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{f0.periodtuningpp}
\alias{f0.periodtuningpp}
\title{Period tuning for Predator-Prey base}
\usage{
f0.periodtuningpp(x1, x2, x3, x4, period)
}
\arguments{
\item{x1}{The growth rate of prey}

\item{x2}{The decay rate of predator}

\item{x3}{The predating effect on prey}

\item{x4}{The predating effecto on predator}

\item{period}{The desired oscilation period}
}
\value{
The solution fitness cost
}
\description{
This function is an example on how EvoPER can be
used for estimating the parameter values in order to produce
oscilations with the desired period. It is not intended to be
used directelly, the provided wrappers should be instead.
}

