% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{f0.rosenbrock2}
\alias{f0.rosenbrock2}
\title{f0.rosenbrock2}
\usage{
f0.rosenbrock2(x1, x2)
}
\arguments{
\item{x1}{Parameter 1}

\item{x2}{Parameter 2}
}
\description{
Two variable Rosenbrock function, where f(1,1) = 0
}

