% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{f0.rosenbrockn}
\alias{f0.rosenbrockn}
\title{f0.rosenbrockn}
\usage{
f0.rosenbrockn(...)
}
\arguments{
\item{...}{The variadic list of function variables.}
}
\value{
The function value
}
\description{
The rosenbrock function of N variables for testing
optimization methods. The global optima for the function is given by
xi = 1, forall i E {1...N}, f(x) = 0.
}
\references{
http://deap.gel.ulaval.ca/doc/dev/api/benchmarks.html
}
