% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{f0.schaffer}
\alias{f0.schaffer}
\title{f0.schaffer}
\usage{
f0.schaffer(...)
}
\arguments{
\item{...}{The variadic list of function variables.}
}
\value{
The function value
}
\description{
The schaffer function of N variables for testing
optimization methods. The global optima for the function is given by
xi = 0, forall i E \{1...N\}, f(x) = 0.
}
\references{
http://deap.gel.ulaval.ca/doc/dev/api/benchmarks.html
}
