% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{f0.schwefel}
\alias{f0.schwefel}
\title{f0.schwefel}
\usage{
f0.schwefel(...)
}
\arguments{
\item{...}{The variadic list of function variables.}
}
\value{
The function value
}
\description{
The schwefel function of N variables for testing optimization
methods. The global optima for the function is given by
xi = 420.96874636, forall i E {1...N}, f(x) = 0. The range of xi is [-500,500]
}
\references{
http://deap.gel.ulaval.ca/doc/dev/api/benchmarks.html
}

