% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{f0.schwefel4}
\alias{f0.schwefel4}
\title{f0.schwefel4}
\usage{
f0.schwefel4(x1, x2, x3, x4)
}
\arguments{
\item{x1}{The first function variable}

\item{x2}{The second function variable}

\item{x3}{The third function variable}

\item{x4}{The fourth function variable}
}
\value{
The function value
}
\description{
The schwefel function of N variables for testing optimization
methods. The global optima for the function is given by
xi = 420.96874636, forall i E {1...N}, f(x) = 0. The range of xi is [-500,500]
}
