% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{f0.test}
\alias{f0.test}
\title{f0.test}
\usage{
f0.test(x1, x2, x3, x4)
}
\arguments{
\item{x1}{Parameter 1}

\item{x2}{Parameter 2}

\item{x3}{Parameter 3}

\item{x4}{Parameter 4}
}
\description{
Simple test function f(1,2,3,4) = 0
}

