% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{f1.cigar}
\alias{f1.cigar}
\title{f1.cigar}
\usage{
f1.cigar(x)
}
\arguments{
\item{x}{The vector of function variables.}
}
\value{
The function value
}
\description{
The Cigar function of N variables for testing optimization
methods. The global optima for the function is given by
xi = 0, forall i E \{1...N\}, f(x) = 0.
}
\references{
http://deap.gel.ulaval.ca/doc/dev/api/benchmarks.html
}
