% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{f1.rastrigin}
\alias{f1.rastrigin}
\title{f1.rastrigin}
\usage{
f1.rastrigin(x)
}
\arguments{
\item{x}{The vector of function parameters}
}
\value{
The function value
}
\description{
The rastrigin function of N variables for testing
optimization methods. The range is xi E [-5.12, 5.12].
The global optima for the function is given by
xi = 0, forall i E \{1...N\}, f(x) = 0.
}
\references{
http://deap.gel.ulaval.ca/doc/dev/api/benchmarks.html#deap.benchmarks.rastrigin
}
