% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{f1.rosenbrock2}
\alias{f1.rosenbrock2}
\title{f1.rosenbrock2}
\usage{
f1.rosenbrock2(x)
}
\arguments{
\item{x}{Parameter vector}
}
\description{
Two variable Rosenbrock function, where f(c(1,1)) = 0
}
