% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{f1.rosenbrockn}
\alias{f1.rosenbrockn}
\title{f1.rosenbrockn}
\usage{
f1.rosenbrockn(x)
}
\arguments{
\item{x}{The vector of function parameters}
}
\value{
The function value
}
\description{
The rosenbrock function of N variables for testing
optimization methods. The global optima for the function is given by
xi = 1, forall i E {1...N}, f(x) = 0.
}
\references{
http://deap.gel.ulaval.ca/doc/dev/api/benchmarks.html
}

