% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{f1.schwefel}
\alias{f1.schwefel}
\title{f1.schwefel}
\usage{
f1.schwefel(x)
}
\arguments{
\item{x}{The vector of function variables.}
}
\value{
The function value
}
\description{
The schwefel function of N variables for testing optimization
methods. The global optima for the function is given by
xi = 420.96874636, forall i E \{1...N\}, f(x) = 0. The range of xi is [-500,500]
}
\references{
http://deap.gel.ulaval.ca/doc/dev/api/benchmarks.html
}
