% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{fixdfcolumns}
\alias{fixdfcolumns}
\title{fixdfcolumns}
\usage{
fixdfcolumns(df, cols = c(), skip = TRUE, type = as.numeric)
}
\arguments{
\item{df}{The data frame.}

\item{cols}{The dataframe columns to be skiped or included.}

\item{skip}{If TRUE the column names in 'cols' are skiped. When FALSE logic is inverted.}

\item{type}{The type for which data frame columns must be converted.}
}
\value{
The data frame with converted column types.
}
\description{
Coerce dataframe columns to a specic type.
}

