% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{generateSolution}
\alias{generateSolution}
\title{generateSolution}
\usage{
generateSolution(parameters, size)
}
\arguments{
\item{parameters}{The Objective Function parameter list}

\item{size}{The solution size}
}
\value{
The solution set
}
\description{
Generates a problema solution using discrete leves
}
