% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{getFitness}
\alias{getFitness}
\title{getFitness}
\usage{
getFitness(S, i = NULL)
}
\arguments{
\item{S}{The solution set}

\item{i}{The fitness index, if null return the whole column.}
}
\value{
The selected fitness entry
}
\description{
Given a set S of N solutions created with sortSolution, this function
returns the solution component fot the best solution.
}

