% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{getSolution}
\alias{getSolution}
\title{getSolution}
\usage{
getSolution(S)
}
\arguments{
\item{S}{The solution set}
}
\value{
The solution set
}
\description{
Given a set S of N solutions created with sortSolution, this function
returns the solution component. A solutions is a set of solutions and their associated
fitness
}

