% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{gm.sd}
\alias{gm.sd}
\title{gm.sd}
\usage{
gm.sd(x, mu = NULL)
}
\arguments{
\item{x}{data}

\item{mu}{The geometric mean. If not provided it is calculated.}
}
\value{
geometric standard deviation for data
}
\description{
Simple implementation for geometric standard deviation
}

