% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{histplothelper}
\alias{histplothelper}
\title{histplothelper}
\usage{
histplothelper(d, x, title = NULL)
}
\arguments{
\item{d}{A data frame.}

\item{x}{A string with the dataframe column name for histogram}

\item{title}{The plot title}
}
\value{
A ggplot2 plot object
}
\description{
Simple helper for ploting histograms
}
