% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{lowerBound}
\alias{lowerBound}
\title{lowerBound}
\usage{
lowerBound(particles, factors)
}
\arguments{
\item{particles}{The particle set}

\item{factors}{the defined range for objective function parameters}
}
\value{
The particle greater than or equal to lower limit
}
\description{
Checks if parameters is greater than the lower bounds
}

