% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{Magnitude}
\alias{Magnitude}
\title{Magnitude}
\usage{
Magnitude(v)
}
\arguments{
\item{v}{The numerical value}
}
\value{
The magnitude order
}
\description{
Calculates the magnitude order for a given value
}
