% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{naiveperiod}
\alias{naiveperiod}
\title{naiveperiod}
\usage{
naiveperiod(d)
}
\arguments{
\item{d}{The data to search period}
}
\value{
A list with the average period and amplitude
}
\description{
A naive approach for finding the period in a series
of data points
}
