% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netlogo-helper.R
\name{NLWrapper.FindJar}
\alias{NLWrapper.FindJar}
\title{NLWrapper.FindJar}
\usage{
NLWrapper.FindJar(path)
}
\arguments{
\item{path}{The base path for searching}
}
\value{
The path for NetLogo jar file
}
\description{
Search for the netlogo jar file on the provided path
}
