% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netlogo-helper.R
\name{NLWrapper.Model}
\alias{NLWrapper.Model}
\title{NLWrapper.Model}
\usage{
NLWrapper.Model(netlogodir, modelfile, dataset, maxtime)
}
\arguments{
\item{netlogodir}{The base path of NetLogo installation}

\item{modelfile}{The absolute path for NetLogo model file}

\item{dataset}{The names of model variables}

\item{maxtime}{The total number of iterations}
}
\description{
This wrapper prepares the environment and instantiates the model
}
\examples{
\dontrun{
   rm(list=ls())
   p<- "C:/Program Files/NetLogo 6.0.4/app"
   output<- c("count sheep", "count wolves")
   m<- file.path(nlpath, "models", "Sample Models", "Biology", model, "Wolf Sheep Predation.nlogo")
   o<- NLWrapper.Model(p, m, output, 150)
}

}
