% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netlogo-helper.R
\name{NLWrapper.SetParameter}
\alias{NLWrapper.SetParameter}
\title{NLWrapper.SetParameter}
\usage{
NLWrapper.SetParameter(obj, parameters)
}
\arguments{
\item{obj}{The object retuned by \link{NLWrapper.Model}}

\item{parameters}{The data frame containing the paramters}
}
\description{
Set parameter values
}
\examples{
\dontrun{
   rm(list=ls())
   p<- "C:/Program Files/NetLogo 6.0.4/app"
   m<- file.path(nlpath, "models", "Sample Models", "Biology", "Wolf Sheep Predation.nlogo")
   o<- NLWrapper.Model(p, m)
}

}
