% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netlogo-helper.R
\name{NLWrapper.SetRandomSeed}
\alias{NLWrapper.SetRandomSeed}
\title{NLWrapper.SetRandomSeed}
\usage{
NLWrapper.SetRandomSeed(obj, seed)
}
\arguments{
\item{obj}{The object retuned by \link{NLWrapper.Model}}

\item{seed}{The new random seed}
}
\description{
Configures the random seed
}
