% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netlogo-helper.R
\name{NLWrapper.Shutdown}
\alias{NLWrapper.Shutdown}
\title{NLWrapper.Shutdown}
\usage{
NLWrapper.Shutdown(obj)
}
\arguments{
\item{obj}{The object retuned by \link{NLWrapper.Model}}
}
\description{
This wrapper terminates RNetLogo execution environment
}
