% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-objective.R
\docType{class}
\name{ObjectiveFunction-class}
\alias{ObjectiveFunction}
\alias{ObjectiveFunction-class}
\title{ObjectiveFunction class}
\description{
The base class for optimization functions.
}
\section{Fields}{

\describe{
\item{\code{object}}{The raw output of objective function}

\item{\code{objective}}{The objective function}

\item{\code{parameters}}{The parameter list for objective function}

\item{\code{value}}{The results from objective function}
}}

