% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\docType{class}
\name{Options-class}
\alias{Options}
\alias{Options-class}
\title{Options}
\description{
The base class for the options for the optimization methods.
}
\section{Fields}{

\describe{
\item{\code{type}}{The configuration type}

\item{\code{container}}{The object holding the configuration otions}
}}

