% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-options.R
\docType{class}
\name{OptionsEES2-class}
\alias{OptionsEES2-class}
\alias{OptionsEES2}
\title{OptionsEES2}
\description{
Options for Serial Dilutions method
}
\section{Fields}{

\describe{
\item{\code{dilutions}}{The desired dilutions}
}}


