% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-options.R
\name{OptionsFactory}
\alias{OptionsFactory}
\title{OptionsFactory}
\usage{
OptionsFactory(type, v = NULL)
}
\arguments{
\item{type}{The metaheuristic method}

\item{v}{The options object}
}
\value{
Options object
}
\description{
Instantiate the Options class required for
the specific metaheuristic method.
}
