% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\docType{class}
\name{OptionsSAA-class}
\alias{OptionsSAA}
\alias{OptionsSAA-class}
\title{OptionsSAA}
\description{
Options for SAA method
}
\section{Fields}{

\describe{
\item{\code{temperature}}{The temperature function}
}}

