% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\docType{class}
\name{OptionsSDA-class}
\alias{OptionsSDA}
\alias{OptionsSDA-class}
\title{OptionsSDA}
\description{
Options for Serial Dilutions method
}
\section{Fields}{

\describe{
\item{\code{dilutions}}{The desired dilutions}
}}

