% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{paramconverter}
\alias{paramconverter}
\title{paramconverter}
\usage{
paramconverter(parameters, discrete, levelz = 5)
}
\arguments{
\item{parameters}{The current parameter set}

\item{discrete}{The desired parameter type}

\item{levelz}{When discrete is true the number of levels to be generated}
}
\value{
The parameter collection casted to desired mode
}
\description{
Convert parameter from continuous to discrete
and vice-versa if needed
}
