% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{partSolutionSpace}
\alias{partSolutionSpace}
\title{partSolutionSpace}
\usage{
partSolutionSpace(parameters, d = 4)
}
\arguments{
\item{parameters}{The Objective Function parameter list}

\item{d}{The partition size. Default value 4.}
}
\value{
A set of solutions
}
\description{
Creates the initial Solution population
taking into account the lower an upper bounds of
provided experiment factors. This method works by
dividing the solution space into partitions of size 'd'
and then creating a full factorial combination of partitions.
}

