% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{predatorprey.plot}
\alias{predatorprey.plot}
\title{predatorprey.plot}
\usage{
predatorprey.plot(x1, x2, x3, x4)
}
\arguments{
\item{x1}{The growth rate of prey}

\item{x2}{The decay rate of predator}

\item{x3}{The predating effect on prey}

\item{x4}{The predating effecto on predator}
}
\value{
An ggplot2 object
}
\description{
Generate a plot for the predator-prey ODE output.
}
\examples{
\dontrun{
 predatorprey.plot(1.351888, 1.439185, 1.337083, 0.9079049)
}

}

