% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{predatorprey.plot1}
\alias{predatorprey.plot1}
\title{predatorprey.plot1}
\usage{
predatorprey.plot1(x, title = NULL)
}
\arguments{
\item{x}{A list containing the values of predator/prey parameters c1, c2, c3 and c4
denoting respectivelly the growth rate of prey, the decay rate of predator,
the predating effect on prey and the predating effect on predator}

\item{title}{The optional plot title. May be omited.}
}
\value{
An ggplot2 object
}
\description{
Simple wrapper for 'predatorprey.plot0' accepting the parameters as a list.
}
\examples{
\dontrun{
 rm(list=ls())
 predatorprey.plot1(v$getBest()[1:4])
}

}

