% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-functions.R
\name{predatorprey}
\alias{predatorprey}
\title{predatorprey}
\usage{
predatorprey(x1, x2, x3, x4)
}
\arguments{
\item{x1}{The growth rate of prey}

\item{x2}{The decay rate of predator}

\item{x3}{The predating effect on prey}

\item{x4}{The predating effecto on predator}
}
\value{
The ODE solution
}
\description{
The solver for Lotka-Volterra differential equation.
}
