% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{pso.best}
\alias{pso.best}
\title{pso.best}
\usage{
pso.best(objective, particles)
}
\arguments{
\item{objective}{The results of evaluating the objective function}

\item{particles}{The particles tested}
}
\value{
The best particle
}
\description{
Search for the best particle solution which minimize
the objective function.
}

