% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-pso.R
\name{pso.chi}
\alias{pso.chi}
\title{pso.chi}
\usage{
pso.chi(phi1, phi2)
}
\arguments{
\item{phi1}{Acceleration coefficient toward the previous best}

\item{phi2}{Acceleration coefficient toward the global best}
}
\value{
The calculated constriction coefficient
}
\description{
Implementation of constriction coefficient
}
