% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-pso.R
\name{pso.lbest}
\alias{pso.lbest}
\title{pso.lbest}
\usage{
pso.lbest(i, pbest, topology)
}
\arguments{
\item{i}{The particle position}

\item{pbest}{The pbest particle collection}

\item{topology}{The desired topology function}
}
\value{
The lbes for i th particle
}
\description{
Finds the lbest for the particle 'i' using the
topology function given by the topology parameter.
}

