% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{pso.lowerBound}
\alias{pso.lowerBound}
\title{lowerBound}
\usage{
pso.lowerBound(particles, factors)
}
\arguments{
\item{particles}{The particle set}

\item{factors}{the defined range for objective function parameters}
}
\value{
The particle greater than or equal to lower limit
}
\description{
Checks if parameters is greater than the lower bounds
}

