% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-pso.R
\name{pso.neighborhood.K2}
\alias{pso.neighborhood.K2}
\title{pso.neighborhood.K2}
\usage{
pso.neighborhood.K2(i, n)
}
\arguments{
\item{i}{The particle position}

\item{n}{the size of particle population}
}
\description{
The neighborhood function for a simple linear topology
where every particle has k = 2 neighbors
}

