% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-pso.R
\name{pso.neighborhood.K4}
\alias{pso.neighborhood.K4}
\title{pso.neighborhood.K4}
\usage{
pso.neighborhood.K4(i, n)
}
\arguments{
\item{i}{The particle position}

\item{n}{the size of particle population}
}
\description{
The von neumann neighborhood function for
a lattice-based topology where every particle
has k = 4 neighbors
}
