% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{pso.neighborhood.KN}
\alias{pso.neighborhood.KN}
\title{pso.neighborhood.KN}
\usage{
pso.neighborhood.KN(i, n)
}
\arguments{
\item{i}{The particle position}

\item{n}{the size of particle population}
}
\description{
Simple helper method for 'gbest' neighborhood
}

