% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{pso.printbest}
\alias{pso.printbest}
\title{pso.printbest}
\usage{
pso.printbest(objective, particles, generation, title)
}
\arguments{
\item{objective}{An instance of ObjectiveFunction (or subclass) class \link{ObjectiveFunction}}

\item{particles}{The current particle population}

\item{generation}{The current generation}

\item{title}{Some informational text to be shown}
}
\description{
Shows the best particle of each of simulated generations
}

