% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{pso.Velocity}
\alias{pso.Velocity}
\title{pso.velocity}
\usage{
pso.Velocity(W = 1, Vi, phi1, phi2, Pi, Pg, Xi)
}
\arguments{
\item{W}{Weight (Inertia weight or constriction coefficient)}

\item{Vi}{Current Velocity vector}

\item{phi1}{Acceleration coefficient toward the previous best}

\item{phi2}{Acceleration coefficient toward the global best}

\item{Pi}{Personal best}

\item{Pg}{Neighborhood best}

\item{Xi}{Particle vector}
}
\value{
Updated velocity
}
\description{
Calculates the PSO Velocity
}

