% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{push}
\alias{push}
\title{push}
\usage{
push(x, v)
}
\arguments{
\item{x}{The collection of elements}

\item{v}{The value to be pushed}
}
\value{
The collection of elements
}
\description{
push an element
}
