% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{rangesearch.pso}
\alias{rangesearch.pso}
\title{rangesearch.pso}
\usage{
rangesearch.pso(aproximations = 10, replications = 4, objective,
  iterations = 30, iwc = 10, N = 10, phi1 = 2, phi2 = 2, W = 0)
}
\arguments{
\item{aproximations}{The number of aproximations}

\item{replications}{The number of repetitions of each aproximation}

\item{objective}{An instance of ObjectiveFunction (or subclass) class \link{ObjectiveFunction}}

\item{iterations}{The total number of interactions}

\item{iwc}{The maximun number of iteractions without changes in the output}

\item{N}{The Particle Swarm size}

\item{phi1}{Acceleration coefficient toward the previous best}

\item{phi2}{Acceleration coefficient toward the global best}

\item{W}{Inertia weight or Constriction coefficient}
}
\description{
This function tries to provide a better approximation to the
best solution when no information is available on the correct range of
input parameters for the objective function. Basically the function search
for the best cost n replications and adjust the range based on that value
of best particles and a 10% of initially provided range.
}

