% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{saa.neighborhood}
\alias{saa.neighborhood}
\title{saa.neighborhood}
\usage{
saa.neighborhood(f, S, d, n)
}
\arguments{
\item{f}{An instance of ObjectiveFunction (or subclass) class \link{ObjectiveFunction}}

\item{S}{The current solution to find a neighbor}

\item{d}{The distance from current solution S \code{distance = (max - min) * d}}

\item{n}{The number of parameters to be perturbed}
}
\value{
The neighbor of solution S
}
\description{
Generates neighbor solutions for simulated annealing
}

