% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-saa.R
\name{saa.neighborhood1}
\alias{saa.neighborhood1}
\title{saa.neighborhood1}
\usage{
saa.neighborhood1(f, S, d)
}
\arguments{
\item{f}{An instance of ObjectiveFunction (or subclass) class \link{ObjectiveFunction}}

\item{S}{The current solution to find a neighbor}

\item{d}{The distance from current solution S \code{distance = (max - min) * d}}
}
\value{
The neighbor of solution of S
}
\description{
Generates neighbor solutions perturbing one parameter from current
solution \code{S} picked randonly.
}

