% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{saa.tbyk}
\alias{saa.tbyk}
\title{saa.tbyk}
\usage{
saa.tbyk(t0, k)
}
\arguments{
\item{t0}{The current temperature}

\item{k}{The annealing value}
}
\value{
The new temperature
}
\description{
Temperature function t/k
}

