% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-saa.R
\name{saa.tcte}
\alias{saa.tcte}
\title{saa.tcte}
\usage{
saa.tcte(t0, k)
}
\arguments{
\item{t0}{The current temperature}

\item{k}{The annealing value}
}
\value{
The new temperature
}
\description{
Temperature function cte * t0
}

