% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-saa.R
\name{saa.texp}
\alias{saa.texp}
\title{saa.texp}
\usage{
saa.texp(t0, k)
}
\arguments{
\item{t0}{The current temperature}

\item{k}{The annealing value}
}
\value{
The new temperature
}
\description{
Temperature function exponential
}
