% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{scatterplotlothelper}
\alias{scatterplotlothelper}
\title{scatterplotlothelper}
\usage{
scatterplotlothelper(d, x, y, z, title = NULL)
}
\arguments{
\item{d}{A data frame.}

\item{x}{A string with the dataframe column name for x axis}

\item{y}{A string with the dataframe column name for y axis}

\item{z}{A string with the dataframe column name for z axis}

\item{title}{The plot title}
}
\value{
A scatter3D plot
}
\description{
Simple helper for ploting 3d scaterplots
}

