% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{searchrow}
\alias{searchrow}
\title{searchrow}
\usage{
searchrow(ddata, value)
}
\arguments{
\item{ddata}{The matrix containing the dataset}

\item{value}{The value to search for}
}
\value{
Boolean TRUE for those indexes matching value
}
\description{
Search for a value value on a matrix
}
