% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{SetLogLevel}
\alias{SetLogLevel}
\title{SetLogLevel}
\usage{
SetLogLevel(level)
}
\arguments{
\item{level}{The log level (ERROR|WARN|INFO|DEBUG)}
}
\description{
Configure the current log level for the package
}
