% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{show.comp1}
\alias{show.comp1}
\title{show.comp1}
\usage{
show.comp1(mydata, what, title = NULL)
}
\arguments{
\item{mydata}{The data generated with 'summarize.comp1'}

\item{what}{The name of variable to plot on 'y' axis}

\item{title}{the plot title}
}
\description{
Generates a barplot comparing the number of evalutions for
algorithms ("saa","pso","acor","ees1").
}
\examples{
\dontrun{
	p.a<- show.comp1(d.cigar4,"evals","(a) Cigar function")
	p.b<- show.comp1(d.schaffer4,"evals","(b) Schafer function")
	p.c<- show.comp1(d.griewank4,"evals","(c) Griewank function")
	p.d<- show.comp1(d.bohachevsky4,"evals","(d) Bohachevsky function")
}

}

