% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{slope}
\alias{slope}
\title{slope}
\usage{
slope(x, y, i)
}
\arguments{
\item{x}{The x vector}

\item{y}{The y vector}

\item{i}{The position}
}
\value{
The slope
}
\description{
Simple function for calculate the slope on the ith element
position
}

