% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{slopes}
\alias{slopes}
\title{slopes}
\usage{
slopes(x, y)
}
\arguments{
\item{x}{The x vector}

\item{y}{The y vector}
}
\value{
A vector with all slopes
}
\description{
Calcule all slopes for the discrete x,y series
}
