% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{sortSolution}
\alias{sortSolution}
\title{sortSolution}
\usage{
sortSolution(s, f)
}
\arguments{
\item{s}{Problem solution}

\item{f}{The function evaluation for s}
}
\description{
Sort solution by its respective fitness
}

