% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiments.R
\name{summarize.comp1}
\alias{summarize.comp1}
\title{summarize.comp1}
\usage{
summarize.comp1(mydata)
}
\arguments{
\item{mydata}{The data frame generated with 'compare.algorithms1'}
}
\value{
The summarized data
}
\description{
Provides as summary with averged values of experimental setup
}

