% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-tabu.R
\name{tabu.getNeighbors}
\alias{tabu.getNeighbors}
\title{tabu.getNeighbors}
\usage{
tabu.getNeighbors(tabu, parameters, solution, size)
}
\arguments{
\item{tabu}{The tabu list}

\item{parameters}{The parameter set}

\item{solution}{The current solution}

\item{size}{The neigborhood size}
}
\value{
The neighbor for solution
}
\description{
create neighbor solutions
}
