% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm-tabu.R
\name{tabu.istabu}
\alias{tabu.istabu}
\title{tabu.istabu}
\usage{
tabu.istabu(tabulist, solution)
}
\arguments{
\item{tabulist}{The matrix of tabu solutions}

\item{solution}{The solution value to be checked}
}
\value{
Boolean TRUE tabulist contains the solution
}
\description{
Check whether a solution is present on tabulist
}
