% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{upperBound}
\alias{upperBound}
\title{upperBound}
\usage{
upperBound(particles, factors)
}
\arguments{
\item{particles}{The particle set}

\item{factors}{the defined range for objective function parameters}
}
\value{
The particle inside the valid upper bound
}
\description{
Checks if parameters is below the upper bounds
}

