% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{xmeanci1}
\alias{xmeanci1}
\title{xmeanci1}
\usage{
xmeanci1(x, alpha = 0.95)
}
\arguments{
\item{x}{The data set for which CI will be calculated}

\item{alpha}{The confidence level. The default value is 0.95 (95\%)}
}
\value{
The confidence interval for the mean calculated using 'boot.ci'
}
\description{
Calculates confidence interval of mean for provided
data with desired confidence level. This functions uses bootstrap
resampling scheme for estimanting the CI.
}
