% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{xmeanci2}
\alias{xmeanci2}
\title{xmeanci2}
\usage{
xmeanci2(x, alpha = 0.95)
}
\arguments{
\item{x}{The data set for which CI will be calculated}

\item{alpha}{The confidence level. The default value is 0.95 (95\%)}
}
\value{
The confidence interval for the mean
}
\description{
Calculates confidence interval of mean for provided
data with desired confidence level.
}
