% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{abm.pso}
\alias{abm.pso}
\title{abm.pso}
\usage{
abm.pso(objective, iterations = 100, N = 16, phi1 = 1.193, phi2 = 1.193,
  W = 0.721, f.neighborhood = pso.neighborhood.KN)
}
\arguments{
\item{objective}{An instance of ObjectiveFunction (or subclass) class \link{ObjectiveFunction}}

\item{iterations}{The total number of interactions}

\item{N}{The Particle Swarm size}

\item{phi1}{Acceleration coefficient toward the previous best}

\item{phi2}{Acceleration coefficient toward the global best}

\item{W}{Inertia weight or Constriction coefficient}

\item{f.neighborhood}{The neigborhood function}
}
\description{
An implementaion of Particle Swarm Optimization
method for parameter estimation of Individual-based models.
}
\examples{
\dontrun{
 f<- RepastFunction$new("c:/usr/models/BactoSim(HaldaneEngine-1.0)","ds::Output",300)

 f$Parameter(name="cyclePoint",min=0,max=90)
 f$Parameter(name="conjugationCost",min=0,max=100)
 f$Parameter(name="pilusExpressionCost",min=0,max=100)
 f$Parameter(name="gamma0",min=1,max=10)

 abm.pso(f, iterations=100)
}

}
\references{
[1] Kennedy, J., & Eberhart, R. (1995). Particle swarm optimization.
In Proceedings of ICNN 95 - International Conference on Neural
Networks (Vol. 4, pp. 1942-1948). IEEE.

[2] Poli, R., Kennedy, J., & Blackwell, T. (2007). Particle swarm optimization.
Swarm Intelligence, 1(1), 33-57.
}

