% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{ees1.explore}
\alias{ees1.explore}
\title{ees1.explore}
\usage{
ees1.explore(s, weight)
}
\arguments{
\item{s}{The Problem solution}

\item{weight}{The exploration intensity}
}
\description{
Explore the solution space on the neighborhood of
solution 's' in order to find a new best.
}

