% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{initSolution}
\alias{initSolution}
\title{initSolution}
\usage{
initSolution(parameters, N = 20)
}
\arguments{
\item{parameters}{The Objective Function parameter list}

\item{N}{The size of Solution population}
}
\value{
A random set of solutions
}
\description{
Creates the initial Solution population
taking into account the lower an upper bounds of
provided experiment factors.
}

