% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netlogo-helper.R
\name{NLWrapper.Run}
\alias{NLWrapper.Run}
\title{NLWrapper.Run}
\usage{
NLWrapper.Run(obj, r = 1, seed = c())
}
\arguments{
\item{obj}{The object retuned by \link{NLWrapper.Model}}

\item{r}{The number of replications}

\item{seed}{The collection of random seeds}
}
\description{
Executes a NetLogo Model using rNetLogo
}
\examples{
\dontrun{
p<- "C:/Program Files/NetLogo 6.1.1/app"
m<- file.path(p, "models", "Sample Models", "Biology", "Wolf Sheep Predation.nlogo")
output<- c("count sheep", "count wolves")
o<- NLWrapper.Model(p, m, output, 150)
v<- NLWrapper.Run(o)
NLWrapper.Shutdown(o)}


}
