% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-options.R
\docType{class}
\name{Options-class}
\alias{Options-class}
\alias{Options}
\title{Options}
\description{
The base class for the options for the optimization metaheuristics
}
\section{Fields}{

\describe{
\item{\code{type}}{The configuration type}

\item{\code{neighborhood}}{The neighborhood function for population methods}

\item{\code{discrete}}{Flag indicating that and specific algorithm is discrete or continuous}

\item{\code{nlevelz}}{Default value for generating parameter levels when range is provided, default value is 5}

\item{\code{container}}{The object holding the configuration otions}
}}


