% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-options.R
\docType{class}
\name{Options-class}
\alias{Options}
\alias{Options-class}
\title{Options}
\description{
The base class for the options for the optimization metaheuristics
}
\section{Fields}{

\describe{
\item{\code{type}}{The configuration type}

\item{\code{neighborhood}}{The neighborhood function for population methods}

\item{\code{container}}{The object holding the configuration otions}
}}

