% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoper.R
\name{xyplothelper}
\alias{xyplothelper}
\title{xyplothelper}
\usage{
xyplothelper(d, x, y, title = NULL)
}
\arguments{
\item{d}{A data frame.}

\item{x}{A string with the dataframe column name for x axis}

\item{y}{A string with the dataframe column name for y axis}

\item{title}{The plot title}
}
\value{
A ggplot2 plot object
}
\description{
Simple helper for ploting xy dispersion points.
}

